Add prod.js for starting monolith on fly
