--- gateway/index.js
@@ -1,3 +1,0 @@
-import { start } from "./server.js";
-
-await start();
