--- gateway/package.json
@@ -1,17 +1,0 @@
-{
-  "name": "@apollosolutions/retail-supergraph-gateway",
-  "version": "1.0.0",
-  "main": "index.js",
-  "private": true,
-  "license": "MIT",
-  "type": "module",
-  "scripts": {
-    "start": "node index.js",
-    "dev": "NODE_ENV=dev nodemon index.js"
-  },
-  "dependencies": {
-    "@apollo/gateway": "^2.2.1",
-    "@apollo/server": "^4.2.2",
-    "graphql": "^16.6.0"
-  }
-}
