+++ package.json
@@ -2,17 +2,23 @@
   "name": "@apollosolutions/retail-supergraph",
   "license": "MIT",
   "type": "module",
-  "workspaces": [
-    "gateway",
-    "subgraphs/*"
-  ],
   "scripts": {
     "dev": "NODE_ENV=dev nodemon dev.js",
-    "start": "node subgraphs.js",
+    "start": "node prod.js",
     "health-check": "curl -H 'content-type: application/json' http://localhost:4000/graphql?query=%7B__typename%7D",
     "test": "npm run test:compile && npm run test:runtime",
     "test:runtime": "start-server-and-test dev http-get://localhost:4000/?query={__typename} health-check",
     "test:compile": "tsc"
+  },
+  "dependencies": {
+    "@apollo/gateway": "^2.2.1",
+    "@apollo/server": "^4.2.2",
+    "@apollo/subgraph": "^2.2.2",
+    "body-parser": "^1.20.1",
+    "cors": "^2.8.5",
+    "express": "^4.18.2",
+    "graphql": "^16.6.0",
+    "uuid": "^9.0.0"
   },
   "devDependencies": {
     "nodemon": "^2.0.20",
