+++ prod.js
@@ -1,0 +1,6 @@
+import { startSubgraphs } from "./subgraphs.js";
+
+(async () => {
+  // start subgraphs in monolith mode
+  await startSubgraphs();
+})();
