+++ subgraphs.js
@@ -2,24 +2,49 @@ import express from 'express';
 import http from 'http';
 import cors from 'cors';
 import bodyParser from 'body-parser';
-import { getSchema as getUsersSchema } from "@apollosolutions/retail-supergraph-users/subgraph.js";
-import {expressMiddleware} from "@apollo/server/express4";
-import {ApolloServer} from "@apollo/server";
+import { getSchema as getCheckoutSchema } from "./subgraphs/checkout/subgraph.js";
+import { getSchema as getDiscoverySchema } from "./subgraphs/discovery/subgraph.js";
+import { getSchema as getInventorySchema } from "./subgraphs/inventory/subgraph.js";
+import { getSchema as getOrdersSchema } from "./subgraphs/orders/subgraph.js";
+import { getSchema as getProductsSchema } from "./subgraphs/products/subgraph.js";
+import { getSchema as getReviewsSchema } from "./subgraphs/reviews/subgraph.js";
+import { getSchema as getShippingSchema } from "./subgraphs/shipping/subgraph.js";
+import { getSchema as getUsersSchema } from "./subgraphs/users/subgraph.js";
+import { expressMiddleware } from "@apollo/server/express4";
+import { ApolloServer } from "@apollo/server";
 import { ApolloServerPluginDrainHttpServer } from '@apollo/server/plugin/drainHttpServer';
-import {start as checkout} from "@apollosolutions/retail-supergraph-checkout/server.js";
-import {start as discovery} from "@apollosolutions/retail-supergraph-discovery/server.js";
-import {start as inventory} from "@apollosolutions/retail-supergraph-inventory/server.js";
-import {start as orders} from "@apollosolutions/retail-supergraph-orders/server.js";
-import {start as products} from "@apollosolutions/retail-supergraph-products/server.js";
-import {start as reviews} from "@apollosolutions/retail-supergraph-reviews/server.js";
-import {start as shipping} from "@apollosolutions/retail-supergraph-shipping/server.js";
-import {start as gateway} from "@apollosolutions/retail-supergraph-gateway/server.js";
 
 export const LOCAL_SUBGRAPH_CONFIG = [
   {
+    name: 'checkout',
+    getSchema: getCheckoutSchema
+  },
+  {
+    name: 'discovery',
+    getSchema: getDiscoverySchema
+  },
+  {
+    name: 'inventory',
+    getSchema: getInventorySchema
+  },
+  {
+    name: 'orders',
+    getSchema: getOrdersSchema
+  },
+  {
+    name: 'products',
+    getSchema: getProductsSchema
+  },
+  {
+    name: 'reviews',
+    getSchema: getReviewsSchema
+  },
+  {
+    name: 'shipping',
+    getSchema: getShippingSchema
+  },
+  {
     name: 'users',
-    port: 4008,
-    url: 'http://localhost:4008/users/graphql',
     getSchema: getUsersSchema
   }
 ];
@@ -27,15 +52,15 @@ export const LOCAL_SUBGRAPH_CONFIG = [
 const getLocalSubgraphConfig = (subgraphName) =>
   LOCAL_SUBGRAPH_CONFIG.find(it => it.name === subgraphName);
 
-export const startSubgraphs = async () => {
+export const startSubgraphs = async (httpPort) => {
   // Create a monolith express app for all subgraphs
   const app = express();
   const httpServer = http.createServer(app);
+  const serverPort = process.env.PORT ?? httpPort;
 
   // Run each subgraph on the same http server, but at different paths
   for (const subgraph of LOCAL_SUBGRAPH_CONFIG) {
     const subgraphConfig = getLocalSubgraphConfig(subgraph.name);
-    const serverPort = process.env.PORT ?? subgraphConfig.port;
     const schema = subgraphConfig.getSchema();
     const server = new ApolloServer({
       schema,
@@ -54,13 +79,9 @@ export const startSubgraphs = async () => {
       })
     );
 
-    await new Promise((resolve) => httpServer.listen({ port: serverPort }, resolve));
-
-    console.log(`${subgraphConfig.name} subgraph running at http://localhost:${serverPort}${path}`);
+    console.log(`Setting up ${subgraphConfig.name} subgraph running at http://localhost:${serverPort}${path}`);
   }
-};
 
-(async () => {
-  // start subgraphs in monolith mode
-  await startSubgraphs();
-})();
+  // Start monolith at given port
+  await new Promise((resolve) => httpServer.listen({ port: serverPort }, resolve));
+};
