--- subgraphs/discovery/fly.toml
@@ -1,43 +1,0 @@
-# fly.toml file generated for apollosolutions--retail-supergraph--discovery on 2022-10-11T12:36:50-07:00
-
-app = "apollosolutions--retail-supergraph--discovery"
-kill_signal = "SIGINT"
-kill_timeout = 5
-processes = []
-
-[build]
-  builder = "heroku/buildpacks:20"
-
-[env]
-  NODE_ENV = "fly"
-  PORT = "8080"
-
-[experimental]
-  allowed_public_ports = []
-  auto_rollback = true
-
-[[services]]
-  http_checks = []
-  internal_port = 8080
-  processes = ["app"]
-  protocol = "tcp"
-  script_checks = []
-  [services.concurrency]
-    hard_limit = 25
-    soft_limit = 20
-    type = "connections"
-
-  [[services.ports]]
-    force_https = true
-    handlers = ["http"]
-    port = 80
-
-  [[services.ports]]
-    handlers = ["tls", "http"]
-    port = 443
-
-  [[services.tcp_checks]]
-    grace_period = "1s"
-    interval = "15s"
-    restart_limit = 0
-    timeout = "2s"
