--- subgraphs/discovery/package.json
@@ -1,17 +1,0 @@
-{
-  "name": "@apollosolutions/retail-supergraph-discovery",
-  "version": "1.0.0",
-  "main": "index.js",
-  "private": true,
-  "license": "MIT",
-  "type": "module",
-  "scripts": {
-    "start": "node index.js",
-    "dev": "NODE_ENV=dev nodemon index.js"
-  },
-  "dependencies": {
-    "@apollo/subgraph": "^2.2.2",
-    "@apollo/server": "^4.2.2",
-    "graphql": "^16.6.0"
-  }
-}
