--- subgraphs/reviews/fly.toml
@@ -1,41 +1,0 @@
-app = "apollosolutions--retail-supergraph--reviews"
-kill_signal = "SIGINT"
-kill_timeout = 5
-processes = []
-
-[build]
-  builder = "heroku/buildpacks:20"
-
-[env]
-  NODE_ENV = "fly"
-  PORT = "8080"
-
-[experimental]
-  allowed_public_ports = []
-  auto_rollback = true
-
-[[services]]
-  http_checks = []
-  internal_port = 8080
-  processes = ["app"]
-  protocol = "tcp"
-  script_checks = []
-  [services.concurrency]
-    hard_limit = 25
-    soft_limit = 20
-    type = "connections"
-
-  [[services.ports]]
-    force_https = true
-    handlers = ["http"]
-    port = 80
-
-  [[services.ports]]
-    handlers = ["tls", "http"]
-    port = 443
-
-  [[services.tcp_checks]]
-    grace_period = "1s"
-    interval = "15s"
-    restart_limit = 0
-    timeout = "2s"
