--- subgraphs/users/package.json
@@ -1,21 +1,0 @@
-{
-  "name": "@apollosolutions/retail-supergraph-users",
-  "version": "1.0.0",
-  "main": "index.js",
-  "private": true,
-  "license": "MIT",
-  "type": "module",
-  "scripts": {
-    "start": "node index.js",
-    "dev": "NODE_ENV=dev nodemon index.js"
-  },
-  "dependencies": {
-    "@apollo/subgraph": "^2.2.2",
-    "@apollo/server": "^4.2.2",
-    "body-parser": "^1.20.1",
-    "cors": "^2.8.5",
-    "express": "^4.18.2",
-    "graphql": "^16.6.0",
-    "uuid": "^9.0.0"
-  }
-}
